+ Mathematics + Stream syntax

+++ Bot.cpp
@@ -3,10 +3,10 @@
 
 Vector <Node *> Bot::nodes = Vector<Node *>();
 Vector <Updatable *> Bot::updatables = Vector<Updatable *>();
-volatile unsigned long Bot::micros = 0;
+volatile float Bot::micros = 0;
 volatile float Bot::millis = 0;
 volatile float Bot::seconds = 0;
-volatile unsigned long Bot::frames = 0;
+volatile float Bot::frames = 0;
